Allow language ngerman (nowadays used new German spelling), remove link to sarovar.org -- it is dead forever.
Nowadays one use ngerman as language option in German LaTeX
documents for the new German spelling. LyX calls the package without
option and nomencl options must be added to the global options.
So, it is not possible to pass the german option to nomencl. At
http://wiki.lyx.org/Tips/Nomenclature
under "Nomenclature and ngerman (German new spelling)" the
problem is described and the applied patch is provided.